requirements.txt
torch
torchvision
torchaudio
onnxruntime
pyyaml
